+++ projects/rocr-runtime/runtime/hsa-runtime/core/inc/runtime.h
@@ -52,6 +52,8 @@
 #include <utility>
 #include <thread>
 #include <shared_mutex>
+#include <random>
+#include <cinttypes>
 #if defined(__linux__)
 #include <sys/un.h>
 #include <xf86drm.h>
@@ -513,6 +515,8 @@ class Runtime {
   bool VirtualMemApiSupported() const { return virtual_mem_api_supported_; }
   bool XnackEnabled() const { return xnack_enabled_; }
   void XnackEnabled(bool enable) { xnack_enabled_ = enable; }
+  bool AqlProfileAvailable() const { return (aqlprofile_lib_ != nullptr); }
+  os::LibHandle AqlProfileLib() const { return aqlprofile_lib_; }
 
   Driver &AgentDriver(DriverType drv_type) {
     auto is_drv_type = [&](const std::unique_ptr<Driver> &d) {
@@ -911,8 +915,8 @@ class Runtime {
 
   std::unique_ptr<AMD::SvmProfileControl> svm_profile_;
 
-  // IPC DMA buf unix domain socket server dmabuf FD passing
-  int ipc_sock_server_fd_;
+  // IPC DMA buf socket server for dmabuf FD passing
+  os::IPCSocket ipc_sock_server_fd_;
   std::map<uint64_t, size_t> ipc_sock_server_conns_;
   std::mutex ipc_sock_server_lock_;
   os::Thread ipc_sock_server_thread_;
@@ -924,6 +928,8 @@ class Runtime {
 
   bool virtual_mem_api_supported_;
   bool xnack_enabled_;
+
+  os::LibHandle aqlprofile_lib_;
 
   typedef void* ThunkHandle;