+++ projects/rocdecode/CHANGELOG.md
@@ -7,7 +7,7 @@ Full documentation for rocDecode is available at [https://rocm.docs.amd.com/proj
 ### Added
 
 * Logging improvement: Added function entry and exit logs (at Info log level).
-* Logging improvement: Added duration to function exit logs. Some optimization.
+* Logging improvement: Added duration to function exit logs and optimized log message formatting to reduce runtime overhead.
 
 ## rocDecode 1.7.0 for ROCm 7.2.1