+++ projects/rocdecode/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Full documentation for rocDecode is available at [https://rocm.docs.amd.com/projects/rocDecode/en/latest/](https://rocm.docs.amd.com/projects/rocDecode/en/latest/)
 
+## (Unreleased) rocDecode 1.8.0
+
+### Added
+
+* Logging improvement: Added function entry and exit logs (at Info log level).
+* Logging improvement: Added duration to function exit logs and optimized log message formatting to reduce runtime overhead.
+* Feature: support for user set output surface format.
+
 ## rocDecode 1.7.0 for ROCm 7.2.1
 
 ### Changed